단일 모듈로 통일
