+++ Dockerfile
@@ -1,0 +1,13 @@
+FROM openjdk:11-jdk
+
+ARG APP_NAME=Shop-Search-System
+ARG VERSION=0.0.1
+
+# JAR_FILE 변수 정의 -> 기본적으로 jar file이 2개이기 때문에 이름을 특정해야함
+ARG JAR_FILE=./build/libs/${APP_NAME}-${VERSION}.jar
+
+# JAR 파일 메인 디렉토리에 복사
+COPY ${JAR_FILE} app.jar
+
+# 시스템 진입점 정의
+ENTRYPOINT ["java","-jar","/app.jar"]
