--- domains/domain/build.gradle
@@ -1,5 +1,0 @@
-
-dependencies {
-    // Elasticsearch
-    implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
-}
