--- domains/domain/src/main/java/team/bakkas/search/DomainApplication.java
@@ -1,12 +1,0 @@
-package team.bakkas.search;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DomainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DomainApplication.class, args);
-    }
-}
