--- domains/domain/src/test/java/team/bakkas/search/DomainApplicationTests.java
@@ -1,13 +1,0 @@
-package team.bakkas.search;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class DomainApplicationTests {
-
-    @Test
-    void contextLoads() {
-
-    }
-}
