--- lambda/functions/build.gradle
@@ -1,30 +1,0 @@
-// gradle에서 사용할 변수 정의
-ext {
-    set('springCloudVersion', "2020.0.2")
-}
-
-dependencies {
-    implementation(project(":domains-domain"))
-
-    implementation 'org.springframework.cloud:spring-cloud-function-adapter-aws'
-    implementation 'com.amazonaws:aws-lambda-java-core:1.2.1'
-    implementation 'com.amazonaws:aws-lambda-java-events:3.8.0'
-
-    // 로컬 테스트를 위한 의존성
-    // implementation 'org.springframework.cloud:spring-cloud-starter-function-web'
-}
-
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
-}
-
-// zip 파일로 압축하기
-task buildzip(type: Zip) {
-    from compileJava
-    from processResources
-    into('lib') {
-        from configurations.runtimeClasspath
-    }
-}
