+++ settings.gradle
@@ -1,37 +1,1 @@
 rootProject.name = 'Shop-Search-System'
-
-["domains", "lambda"].each {
-
-    def compDir = new File(rootDir, it)
-    if(!compDir.exists()){
-        compDir.mkdirs()
-    }
-
-    compDir.eachDir {subDir ->
-
-        def gradleFile = new File(subDir.absolutePath, "build.gradle")
-        if(!gradleFile.exists()){
-            gradleFile.text =
-                    """
-                    dependencies {
-                    }
-                    """.stripIndent(20)
-        }
-
-        [
-                "src/main/java/team/bakkas/search",
-                "src/main/resources",
-                "src/test/java/team/bakkas/search",
-                "src/test/resources"
-        ].each {srcDir->
-            def srcFolder = new File(subDir.absolutePath, srcDir)
-            if(!srcFolder.exists()){
-                srcFolder.mkdirs()
-            }
-        }
-
-        def projectName = ":${it}-${subDir.name}";
-        include projectName
-        project(projectName).projectDir = subDir
-    }
-}
