+++ src/main/java/team/bakkas/search/SearchApplication.java
@@ -3,10 +3,11 @@ package team.bakkas.search;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
-public class FunctionsApplication {
+@SpringBootApplication(scanBasePackages = "team.bakkas")
+public class SearchApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(FunctionsApplication.class, args);
+        SpringApplication.run(SearchApplication.class, args);
     }
+
 }
