+++ src/main/java/team/bakkas/search/handler/LambdaHandler.java
