+++ src/main/java/team/bakkas/search/shop/domain/dto/request/CategoryWithInSearchRequest.java
