+++ src/main/java/team/bakkas/search/shop/domain/dto/request/ShopNameWithInSearchRequest.java
