+++ src/main/java/team/bakkas/search/shop/domain/dto/request/WithInSearchRequest.java
