+++ src/main/java/team/bakkas/search/shop/domain/persist/ShopRepository.java
