+++ src/main/java/team/bakkas/search/shop/domain/vo/DetailCategory.java
