+++ src/main/resources/application-local.yml
@@ -1,0 +1,2 @@
+elasticsearch:
+  host-and-port: localhost:9200
