+++ src/main/resources/application-prod.yml
@@ -1,0 +1,2 @@
+elasticsearch:
+  host-and-port: 43.200.207.212:9200
