+++ src/test/java/team/bakkas/search/SearchApplicationTests.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class FunctionsApplicationTests {
+class SearchApplicationTests {
 
     @Test
     void contextLoads() {
-
     }
+
 }
